backend/requirements.txt
requests
beautifulsoup4
keras
mnist
tensorflow
flake8
black
Flask
Pillow
